Rename semaphore module to semaphore_ll (low level)
