+++ tokio/src/sync/mod.rs
@@ -26,7 +26,7 @@ cfg_sync! {
 
     pub mod oneshot;
 
-    pub(crate) mod semaphore;
+    pub(crate) mod semaphore_ll;
 
     mod task;
     pub(crate) use task::AtomicWaker;
@@ -48,7 +48,7 @@ cfg_not_sync! {
 
     cfg_signal! {
         pub(crate) mod mpsc;
-        pub(crate) mod semaphore;
+        pub(crate) mod semaphore_ll;
     }
 }
